Add date range model #JC19-353
